Change place of calculation for timeincrement
Timeincrement will now be caclculated in class BaseModel instead
of class EnergySystem
